Change divide_by_columns method to handle DatasetDict
